share util functions between ibtracs and ecmwf
